requirements.txt
CairoSVG==2.2.1

Fire==0.1.3
Flask==1.0.2

iPython==7.2.0
Jinja2==2.10
libsass==0.17.0

MatplotLib==2.2.3
Reportlab==3.5.13

Numpy==1.16.0

PDF2Image==1.4.0
PDFkit==0.6.1
Pillow==5.4.1
PyPDF2==1.26.0
Python-dateutil==2.7.5

Wand==0.5.0
WeasyPrint==44
